Version 0.0.4 Finalization
