+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 
 All notable changes to the "remix-forge" extension will be documented in this file. 
 
+## [0.0.4]
+
+### Added
+- Ability to convert v1 route convention to v2 route convention
+- Ability to change runtime dependencies
+
+### Changed
+- Updated readme
+- Moved the commands to a separate context menu section
+
 ## [0.0.3]
 
 - Readme update
